requirements.txt
requests
bottle
waitress
fluent-logger
jaeger-client
opentracing
redis
contextvars
kafka-python
asyncio-nats-client
